restructure as per comments
